+++ food-order-app-sb-jpa-mysql/src/main/java/io/mk/foodorder/entity/item/ItemTaxGroup.java
@@ -18,5 +18,5 @@ public class ItemTaxGroup {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
-	private Float taxPercent;
+	private Integer taxPercent;
 }
